log-consumer/requirements.txt
kafka-python==2.0.2
prometheus_client==0.16.0
